requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.1
email-validator==2.2.0
httpx==0.28.0
PyJWT==2.10.0
python-dotenv==1.0.0
cryptography==44.0.0
python-multipart==0.0.12
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
